Updated tsx file
