+++ src/Grid/Grid.tsx
@@ -15,21 +15,7 @@ const Grid: React.FC<GridProps> = ({ columns, gap = 10, children }) => {
   return <div className="grid" style={gridStyles}>{children}</div>;
 };
 
-const App: React.FC = () => {
-  return (
-    <div className="App">
-      <Grid columns={3} gap={20}>
-        <div className="grid-item">Item 1</div>
-        <div className="grid-item">Item 2</div>
-        <div className="grid-item">Item 3</div>
-        <div className="grid-item">Item 4</div>
-        <div className="grid-item">Item 5</div>
-        <div className="grid-item">Item 6</div>
-      </Grid>
-    </div>
-  );
-};
+export default Grid;
 
-export default App;
 
 /* we have the Grid component implementation followed by the App component where the Grid component is used */
